chore: standardise plugin and zod schema file naming conventions
