+++ typescript/src/plugins/core-token-plugin/index.ts
@@ -2,24 +2,24 @@ import { Context } from '@/shared';
 import { Plugin } from '@/shared/plugin';
 import airdropFungibleToken, {
   AIRDROP_FUNGIBLE_TOKEN_TOOL,
-} from '@/plugins/core-hts-plugin/tools/fungible-token/airdrop-fungible-token';
+} from '@/plugins/core-token-plugin/tools/fungible-token/airdrop-fungible-token';
 import createFungibleTokenTool, {
   CREATE_FUNGIBLE_TOKEN_TOOL,
-} from '@/plugins/core-hts-plugin/tools/fungible-token/create-fungible-token';
+} from '@/plugins/core-token-plugin/tools/fungible-token/create-fungible-token';
 import mintFungibleTokenTool, {
   MINT_FUNGIBLE_TOKEN_TOOL,
-} from '@/plugins/core-hts-plugin/tools/fungible-token/mint-fungible-token';
+} from '@/plugins/core-token-plugin/tools/fungible-token/mint-fungible-token';
 import createNonFungibleTokenTool, {
   CREATE_NON_FUNGIBLE_TOKEN_TOOL,
-} from '@/plugins/core-hts-plugin/tools/non-fungible-token/create-non-fungible-token';
+} from '@/plugins/core-token-plugin/tools/non-fungible-token/create-non-fungible-token';
 import mintNonFungibleTokenTool, {
   MINT_NON_FUNGIBLE_TOKEN_TOOL,
-} from '@/plugins/core-hts-plugin/tools/non-fungible-token/mint-non-fungible-token';
+} from '@/plugins/core-token-plugin/tools/non-fungible-token/mint-non-fungible-token';
 
-export const coreHTSPlugin: Plugin = {
-  name: 'core-hts-plugin',
+export const coreTokenPlugin: Plugin = {
+  name: 'core-token-plugin',
   version: '1.0.0',
-  description: 'A plugin for the Hedera Transaction Service',
+  description: 'A plugin for the Hedera Token Service',
   tools: (context: Context) => {
     return [
       createFungibleTokenTool(context),
@@ -32,7 +32,7 @@ export const coreHTSPlugin: Plugin = {
 };
 
 // Export tool names as an object for destructuring
-export const coreHTSPluginToolNames = {
+export const coreTokenPluginToolNames = {
   AIRDROP_FUNGIBLE_TOKEN_TOOL,
   CREATE_FUNGIBLE_TOKEN_TOOL,
   MINT_FUNGIBLE_TOKEN_TOOL,
@@ -40,4 +40,4 @@ export const coreHTSPluginToolNames = {
   MINT_NON_FUNGIBLE_TOKEN_TOOL,
 } as const;
 
-export default { coreHTSPlugin, coreHTSPluginToolNames };
+export default { coreTokenPlugin, coreTokenPluginToolNames };
