+++ typescript/src/plugins/core-token-plugin/tools/non-fungible-token/mint-non-fungible-token.ts
@@ -4,7 +4,7 @@ import type { Tool } from '@/shared/tools';
 import HederaParameterNormaliser from '@/shared/hedera-utils/hedera-parameter-normaliser';
 import { Client } from '@hashgraph/sdk';
 import { handleTransaction, RawTransactionResponse } from '@/shared/strategies/tx-mode-strategy';
-import { mintNonFungibleTokenParameters } from '@/shared/parameter-schemas/hts.zod';
+import { mintNonFungibleTokenParameters } from '@/shared/parameter-schemas/token.zod';
 import HederaBuilder from '@/shared/hedera-utils/hedera-builder';
 import { PromptGenerator } from '@/shared/utils/prompt-generator';
