+++ typescript/src/plugins/index.ts
@@ -1,15 +1,15 @@
-import { coreHTSPlugin, coreHTSPluginToolNames } from './core-hts-plugin';
+import { coreTokenPlugin, coreTokenPluginToolNames } from './core-token-plugin';
 import { coreAccountPlugin, coreAccountPluginToolNames } from './core-account-plugin';
 import { coreConsensusPlugin, coreConsensusPluginToolNames } from './core-consensus-plugin';
 import { coreQueriesPlugin, coreQueriesPluginToolNames } from './core-queries-plugin';
 import { coreEVMPlugin, coreEVMPluginToolNames } from './core-evm-plugin';
 
 export {
-  coreHTSPlugin,
+  coreTokenPlugin,
   coreAccountPlugin,
   coreConsensusPlugin,
   coreQueriesPlugin,
-  coreHTSPluginToolNames,
+  coreTokenPluginToolNames,
   coreAccountPluginToolNames,
   coreConsensusPluginToolNames,
   coreQueriesPluginToolNames,
