+++ typescript/src/shared/hedera-utils/hedera-parameter-normaliser.ts
@@ -8,19 +8,18 @@ import {
   createNonFungibleTokenParametersNormalised,
   mintFungibleTokenParameters,
   mintNonFungibleTokenParameters,
-} from '@/shared/parameter-schemas/hts.zod';
+} from '@/shared/parameter-schemas/token.zod';
 import {
   deleteAccountParameters,
   deleteAccountParametersNormalised,
   transferHbarParameters,
-  transferHbarParameters,
   updateAccountParameters,
   updateAccountParametersNormalised,
-} from '@/shared/parameter-schemas/has.zod';
+} from '@/shared/parameter-schemas/account.zod';
 import {
   createTopicParameters,
   createTopicParametersNormalised,
-} from '@/shared/parameter-schemas/hcs.zod';
+} from '@/shared/parameter-schemas/consensus.zod';
 
 import { AccountId, Client, Hbar, PublicKey, TokenSupplyType, TokenType } from '@hashgraph/sdk';
 import { Context } from '@/shared/configuration';
